+++ src/browsersession/browser-session-info.ts
@@ -1,0 +1,8 @@
+import {BrowserSessionId, DateTime, DeviceId} from "../types";
+
+export interface BrowserSessionInfo {
+    id: BrowserSessionId,
+    optStart?: DateTime,
+    optEnd?: DateTime,
+    optDeviceId?: DeviceId
+}