+++ src/device/CapabilityConstants.ts
@@ -1,4 +1,6 @@
 export class CapabilityConstants {
     public static readonly SIMULATE_CAMERA = "sensorsimulation.simulateCamera";
     public static readonly MEDIA_SETTINGS = "settings.media";
+    public static readonly SIMULATE_BIOMETRICS = "sensorsimulation.simulateBiometrics";
+    public static readonly ACCEPT_BIOMETRICS = "sensorsimulation.acceptBiometricAuthentication";
 }